+++ www/Repro.js
@@ -11,31 +11,62 @@ Repro.prototype.setLogLevel = function(logLevel, successCallback, errorCallback)
 };
 
 Repro.prototype.startRecording = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "startRecording");
+    exec(successCallback, errorCallback, "Repro", "startRecording", []);
 };
 
 Repro.prototype.stopRecording = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "stopRecording");
+    exec(successCallback, errorCallback, "Repro", "stopRecording", []);
 };
 
 Repro.prototype.pauseRecording = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "pauseRecording");
+    exec(successCallback, errorCallback, "Repro", "pauseRecording", []);
 };
 
 Repro.prototype.resumeRecording = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "resumeRecording");
+    exec(successCallback, errorCallback, "Repro", "resumeRecording", []);
 };
 
-Repro.prototype.maskWithRect = function(x, y, width, height, password, successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "maskWithRect", [x, y, width, height, password]);
+Repro.prototype.maskWithRect = function(key, x, y, width, height, successCallback, errorCallback) {
+    exec(successCallback, errorCallback, "Repro", "maskWithRect", [key, x, y, width, height]);
 };
 
-Repro.prototype.maskFullScreen = function(password, successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "maskFullScreen", [password]);
+// this object holds key for full-screen mask and orientationchange event listener
+Repro.prototype.orientationChangeListener = function(key) {
+    this.key = key;
+    var _key = key;
+    this.listener = function() {
+        exec(null, null, "Repro", "maskFullScreen", [_key]);
+    };
 };
 
-Repro.prototype.unmask = function(password, successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "unmask", [password]);
+// array of orientationChangeListener
+Repro.prototype.orientationChangeListeners = [];
+
+Repro.prototype.maskFullScreen = function(key, successCallback, errorCallback) {
+
+    // add new orientationChangeListener if it's not exists.
+    var found = this.orientationChangeListeners.some(function(l) {return key === l.key});
+    if (!found) {
+        var l = new this.orientationChangeListener(key);
+        window.addEventListener('orientationchange', l.listener);
+        this.orientationChangeListeners.push(l);
+    }
+
+    exec(successCallback, errorCallback, "Repro", "maskFullScreen", [key]);
+};
+
+Repro.prototype.unmask = function(key, successCallback, errorCallback) {
+
+    // remove orientationChangeListener which key euqals 'key'
+    this.orientationChangeListeners = this.orientationChangeListeners.filter(function(l) {
+        if (key === l.key) {
+            window.removeEventListener('orientationchange', l.listener);
+            return false;
+        }
+        return true;
+    });
+
+    exec(successCallback, errorCallback, "Repro", "unmask", [key]);
 };
 
 Repro.prototype.setUserID = function(userId, successCallback, errorCallback) {
@@ -64,16 +95,16 @@ Repro.prototype.setPushDeviceToken = function(deviceToken, successCallback, erro
     exec(successCallback, errorCallback, "Repro", "setPushDeviceToken", [deviceToken]);
 };
 
+Repro.prototype.enablePushNotification = function(senderId, successCallback, errorCallback) {
+    exec(successCallback, errorCallback, "Repro", "enablePushNotification", [senderId]);
+};
+
 Repro.prototype.showInAppMessage = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "showInAppMessage");
+    exec(successCallback, errorCallback, "Repro", "showInAppMessage", []);
 };
 
 Repro.prototype.disableInAppMessageOnActive = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "disableInAppMessageOnActive");
-};
-
-Repro.prototype.survey = function(successCallback, errorCallback) {
-    exec(successCallback, errorCallback, "Repro", "survey");
+    exec(successCallback, errorCallback, "Repro", "disableInAppMessageOnActive", []);
 };
 
 module.exports = new Repro();